+++ src/test/java/janggi/domain/piece/ElephantTest.java
@@ -3,7 +3,6 @@ package janggi.domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import janggi.domain.Position;
 import janggi.domain.Side;
 import java.util.List;
 import java.util.stream.Stream;
